Migrate sanitizeSMILES test to test_utils.py and clean up sanitization and test to use explicit-hydrogen canonical isomeric SMILES
